Added Save Data To Prevent Load Crash
Empty Save Data Slots Cause IronSeed to Crash to Desktop.  The included
Save States in this commit are from the ironseed.net distributable
version of ironSeed, and are intended for testing purposes.  As this is
a code repository, and being able to quickly test game elements is
important, this seems like the best solution at the moment.  In a final
distribution replacing the contents of all the save slots with "Quick
Start" versions would be prudent.