Save Game Fix
The DOS version of IronSeed used all CAPS for its naming convention when
saving games.  As Linux is case sensitive the load and save routines are
a bit of a pain when moving over native save files.  This alteration
saves the end user the hassle of changing the name of every file in
every save directory to lower case.