All Graphics Assets in PNG Format
I am adding all of the graphics assets from a different GitHub
repository of Ironseed.  Once a converter to the IronSeed native cpr
format is found I'll be able to update these accordingly, or code the
Pascal libraries to load these natively.  Either way, this will allow
graphics assets to be changed in the future, and having these assets
present will help.