Data Generators
These files are from an alternate Ironseed Repository, and appear to the
data generators and text files for conversations, items, artifacts,
weapons and so on.  The Pascal files have yet to be converted in this
initial commit, but will be as I go through them.  It is more important
to have this available first, and then work upon it.